Fix type casting in zeroing out matrix row to ensure proper index comparison

+++ kratos/solving_strategies/builder_and_solvers/residualbased_block_builder_and_solver.h
@@ -893,7 +893,7 @@ public:
             if (k_factor == 0.0) {
                 // Zero out the whole row, except the diagonal
                 for (std::size_t j = col_begin; j < col_end; ++j)
-                    if (Acol_indices[j] != static_cast<long int>(Index) )
+                    if (static_cast<std::size_t>(Acol_indices[j]) != Index)
                         Avalues[j] = 0.0;
 
                 // Zero out the RHS
